+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/crypto/MXCryptoConfig.kt
@@ -21,7 +21,7 @@ package im.vector.matrix.android.internal.crypto
  */
 data class MXCryptoConfig(
         // Tell whether the encryption of the event content is enabled for the invited members.
-        // By default, we encrypt messages only for the joined members.
-        // The encryption for the invited members will be blocked if the history visibility is "joined".
-        var enableEncryptionForInvitedMembers: Boolean = false
+        // SDK clients can disable this by settings it to false.
+        // Note that the encryption for the invited members will be blocked if the history visibility is "joined".
+        var enableEncryptionForInvitedMembers: Boolean = true
 )